iperf_vsock: Unlink UNIX domain socket path only in server mode
Delete the UNIX domain socket only when the socket is used for
listening, not for connecting.

This is needed to not get the following error when iperf is running in
server mode:

"iperf3: error - unable to start listener for connections: Address
already in use"

+++ src/iperf_vsock.c
@@ -141,7 +141,8 @@ vsock_sockaddr(const char *cid_str, int port, int listen, socklen_t *len)
 	}
 
 	/* AF_UNIX path is not removed on socket close */
-	(void)unlink(sun->sun_path);
+	if (listen)
+		(void)unlink(sun->sun_path);
 
 	return (struct sockaddr *)sun;
 }
